refactor: [4단계] 패키지 이름 변경
