+++ src/main/kotlin/carrace/logic/car/CarName.kt
@@ -1,4 +1,4 @@
-package car_race.logic.car
+package carrace.logic.car
 
 data class CarName(
     val name: String
