+++ src/main/kotlin/carrace/logic/car/CarNames.kt
@@ -1,4 +1,4 @@
-package car_race.logic.car
+package carrace.logic.car
 
 data class CarNames(val names: List<CarName>) {
     constructor(input: String) : this(input.split(",").map(::CarName))
