+++ src/main/kotlin/carrace/logic/car/Cars.kt
@@ -1,6 +1,6 @@
-package car_race.logic.car
+package carrace.logic.car
 
-import car_race.logic.system.MovingSystem
+import carrace.logic.system.MovingSystem
 
 data class Cars(
     val cars: List<Car>
